rename dashboard to catalogue
