+++ app/app/src/main/java/com/ar/enbaldeapp/ui/catalogue/CatalogueViewModel.java
@@ -1,14 +1,14 @@
-package com.ar.enbaldeapp.ui.dashboard;
+package com.ar.enbaldeapp.ui.catalogue;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class DashboardViewModel extends ViewModel {
+public class CatalogueViewModel extends ViewModel {
 
     private final MutableLiveData<String> mText;
 
-    public DashboardViewModel() {
+    public CatalogueViewModel() {
         mText = new MutableLiveData<>();
         mText.setValue("This is dashboard fragment");
     }
