+++ app/app/src/main/res/layout/fragment_dashboard.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.dashboard.DashboardFragment">
+    tools:context=".ui.catalogue.CatalogueFragment">
 
     <TextView
         android:id="@+id/text_dashboard"
