+++ app/app/src/main/res/navigation/mobile_navigation.xml
@@ -12,8 +12,8 @@
         tools:layout="@layout/fragment_home" />
 
     <fragment
-        android:id="@+id/navigation_dashboard"
-        android:name="com.ar.enbaldeapp.ui.dashboard.DashboardFragment"
+        android:id="@+id/navigation_catalogue"
+        android:name="com.ar.enbaldeapp.ui.catalogue.CatalogueFragment"
         android:label="@string/title_dashboard"
         tools:layout="@layout/fragment_dashboard" />
